Super quick fix/hack for background images borked on Chrome Android
For some weird reason, background-image scaling is borked on Chrome
Android. Seems to happen for the JPG image but NOT the PNG image.
When I change the image source to PNG, everything *works fine*.

Current hack is, instead of just pointing to the larger-file-size PNG,
I create a WebP image from the JPG, which seems to NOT trigger the bug.
Added a WebP detection as well, hopefully will run *before* the JSON
loads.